Migrate lexer from SLY to PLY
This commit migrates the lexer from SLY to PLY by the same author.
The reason for this migration is to gain the ability to do conditional
lexing. Even though SLY is more modern than PLY and has much cleaner
syntax it does not yet support conditional lexing. Conditional lexing
seems to be essential for tokenizing PromQL, strictly as Prometheus'
source code itself does it because of tokens such as DURATION and
for example the differences between IDENTIFIER and METRIC_IDENTIFIER.

This commit also removes the SLY based stub for a parser. It will
be added back in when the lexer is complete.